fix: move access_token tokenization to group_vars (#23)
This pull request introduces updates to the GitHub secrets management
and Ansible configuration for the `helheim_cluster`. The changes
primarily focus on replacing the use of `github_token` with
`access_token` and improving the organization of environment-specific
variables.

### Updates to secrets management:

*
[`.github/workflows/cd-kubernetes-initialize.yml`](diffhunk://#diff-9e0b1332ad1498414505be699f60a085004d1ef74748172457a3c4d155bfafecL28-R29):
Added `ACCESS_TOKEN` to the `TOKENS` secret configuration to support the
updated authentication flow.

### Updates to Ansible configuration:

*
[`src/ansible/inventories/group_vars/helheim_cluster.yml`](diffhunk://#diff-1c53efe9ab014c07a77b79980f44a1691cfe4e5eb12060c224ba1c6c3ebc44a2R2-R3):
Added `github_env` and `github_repo` variables and included
`access_token` in the configuration to align with the new secrets
management approach.
[[1]](diffhunk://#diff-1c53efe9ab014c07a77b79980f44a1691cfe4e5eb12060c224ba1c6c3ebc44a2R2-R3)
[[2]](diffhunk://#diff-1c53efe9ab014c07a77b79980f44a1691cfe4e5eb12060c224ba1c6c3ebc44a2R21)
*
[`src/ansible/roles/kubernetes_leader/tasks/install-k3s.yaml`](diffhunk://#diff-3b0e45f2d3eb69e77b26ccc558e06c7b5b43acb5f943330b27813272aec98f74L61-R61):
Updated the task for writing the GitHub token to a file to use
`access_token` instead of `github_token`.
*
[`src/ansible/roles/kubernetes_leader/defaults/main.yml`](diffhunk://#diff-808d55f47f5eb814161c4fbf4ca29a2cc7b38ef85347c67af942b98208b80e2eL1-L4):
Removed redundant `github_token`, `github_env`, and `github_repo`
variables from the role defaults, as they are now defined elsewhere.