feat: implement nginx gateway manifests for bootstrapping (#30)
This pull request introduces a new `Gateway` resource for the NGINX
gateway and an ArgoCD application configuration to manage it. These
changes enhance the deployment and management of the NGINX gateway in
the Kubernetes cluster.

### NGINX Gateway Configuration:

* Added a new `Gateway` resource in `gateway.yml` to define an HTTPS
listener for the NGINX gateway. This configuration includes TLS
termination with a certificate managed by cert-manager and supports
wildcard hostnames.
(`src/kubernetes/core/nginx-gateway/manifests/gateway.yml`,
[src/kubernetes/core/nginx-gateway/manifests/gateway.ymlR1-R24](diffhunk://#diff-083daf9ba2a633b4a18191cd42752746f05554eddbb7da1313bda47904c378b8R1-R24))

### ArgoCD Application Management:

* Introduced an ArgoCD `Application` resource in `nginx-gateway.yml` to
automate the deployment of the NGINX gateway. This configuration
supports multiple sources, including a Helm chart and a Git repository,
and enables automated synchronization with self-healing and pruning.
(`src/kubernetes/core/nginx-gateway/nginx-gateway.yml`,
[src/kubernetes/core/nginx-gateway/nginx-gateway.ymlR1-R29](diffhunk://#diff-af6eb2591c47fe12de0c4f817d3656c023dfd9baef5efa1d202250a5643da097R1-R29))